+++ src/components/Card/CardExpansion.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-type CardExpansionProps = {};
+export type CardExpansionProps = {};
 
 export const CardExpansion: React.FC<CardExpansionProps> = ({ children }) => {
 	return <>{children}</>;
