+++ src/components/CheckboxGroup/CheckboxGroup.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { FieldSetProps } from '..';
-import { FormFieldSetProps, FormFieldSet } from '../Form_New';
+import { FormFieldSetProps, FormFieldSet } from '../Form';
 import { FieldSet } from '../FieldSet/FieldSet';
 import cx from 'classnames';
 
@@ -35,7 +35,7 @@ type InternalCheckboxGroupProps = {
  */
 export type CheckboxGroupProps<
 	TFieldSetProps extends FieldSetProps | FormFieldSetProps<any>
-> = InternalCheckboxGroupProps &
+	> = InternalCheckboxGroupProps &
 	(TFieldSetProps extends FormFieldSetProps<infer T>
 		? { useFormFieldSet: true } & FormFieldSetProps<T>
 		: FieldSetProps);
@@ -100,7 +100,7 @@ export const CheckboxGroup = <
 const InternalCheckboxGroup: React.FC<InternalCheckboxGroupProps & { id: string }> = ({
 	id,
 	options,
-	onChange = () => {},
+	onChange = () => { },
 	defaultValue = [],
 }) => {
 	const selectedItemsOnInput = Array.isArray(defaultValue) ? defaultValue : [defaultValue];
