--- src/components/ChoiceList/ChoiceListExpansion.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-type ChoiceListExpansionProps = {
-	showOnValue: string;
-};
-
-export const ChoiceListExpansion: React.FC<ChoiceListExpansionProps> = ({ children }) => {
-	return <>{children}</>;
-};
