+++ src/components/Form/FormSubmitButton.tsx
@@ -4,14 +4,15 @@ import { ButtonProps } from '../Button/Button';
 
 type FormSumbitButtonProps = {
 	text: string;
-} & ButtonProps;
+} & Pick<ButtonProps, Exclude<keyof ButtonProps, 'text'>>;
 
 /**
- * Component for adding a form submit button
+ * Component for adding a form submit button,
+ * to be used with generic Form
  * @param props
  */
-const FormSubmitButton: React.FC<FormSumbitButtonProps> = (props) => {
-	return <Button {...props} text={props.text} onClick="submit" />;
+const FormSubmitButton: React.FC<FormSumbitButtonProps> = ({ text, ...props }) => {
+	return <Button {...props} text={text} onClick="submit" />;
 };
 
 export default FormSubmitButton;
