--- src/components/FormComponents/FormFieldStatus.ts
@@ -1,5 +1,0 @@
-import { FormStatusProps } from '../FormStatus/FormStatus';
-
-export type FormFieldStatusProps = {
-	status?: FormStatusProps;
-};
