+++ src/components/Header/Header.tsx
@@ -134,8 +134,8 @@ class HeaderWithoutRouter extends React.Component<HeaderProps & RouteComponentPr
 									{userFirstName ? (
 										<span>Hi, {userFirstName}.</span>
 									) : (
-										<Link to={loginPath}>Sign in</Link>
-									)}
+											<Link to={loginPath}>Sign in</Link>
+										)}
 								</div>
 							</div>
 						</div>
