+++ src/components/Header/NavItem.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-type NavItemType = 'primary' | 'secondary';
+export type NavItemType = 'primary' | 'secondary';
 
 export type NavItemProps = {
 	title: string;
